5_summarizer/requirements.txt
streamlit
PyPDF2
openai==1.90.0
python-dotenv
nltk
streamlit-mic-recorder
pydub
chromadb
sentence-transformers
torch
